requirements.txt
tiktoken
langchain
PyPDF2
python-dotenv
streamlit
faiss-cpu
streamlit-extras
altair
openai
